+++ server/models/tables/category.js
@@ -1,53 +1,46 @@
 module.exports = function (sequelize, DataTypes) {
-  return sequelize.define(
-    "category",
-    {
-      id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        primaryKey: true,
-        autoIncrement: true,
-        unique: true,
+  return sequelize.define('category', {
+    mainId: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      primaryKey: true,
+      references: {
+        model: 'category_main',
+        key: 'id'
       },
-      main_id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        references: {
-          model: "category_main",
-          key: "id",
-        },
-      },
-      sub_id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        references: {
-          model: "category_sub",
-          key: "id",
-        },
-      },
+      field: "main_id"
     },
-    {
-      sequelize,
-      tableName: "category",
-      timestamps: false,
-      indexes: [
-        {
-          name: "PRIMARY",
-          unique: true,
-          using: "BTREE",
-          fields: [{ name: "id" }],
-        },
-        {
-          name: "main_id",
-          using: "BTREE",
-          fields: [{ name: "main_id" }],
-        },
-        {
-          name: "sub_id",
-          using: "BTREE",
-          fields: [{ name: "sub_id" }],
-        },
-      ],
+    subId: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      primaryKey: true,
+      references: {
+        model: 'category_sub',
+        key: 'id'
+      },
+      field: "sub_id"
     }
-  );
+  }, {
+    sequelize,
+    tableName: 'category',
+    timestamps: false,
+    indexes: [
+      {
+        name: "PRIMARY",
+        unique: true,
+        using: "BTREE",
+        fields: [
+          { name: "main_id" },
+          { name: "sub_id" },
+        ]
+      },
+      {
+        name: "category_ibfk_2_idx",
+        using: "BTREE",
+        fields: [
+          { name: "sub_id" },
+        ]
+      },
+    ]
+  });
 };
