+++ server/models/tables/categoryMain.js
@@ -1,30 +1,44 @@
 module.exports = function (sequelize, DataTypes) {
-  return sequelize.define(
-    "categoryMain",
-    {
-      id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        primaryKey: true,
-        unique: true,
-      },
-      name: {
-        type: DataTypes.STRING(20),
-        allowNull: false,
-      },
+  return sequelize.define('categoryMain', {
+    id: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      primaryKey: true
     },
-    {
-      sequelize,
-      tableName: "category_main",
-      timestamps: false,
-      indexes: [
-        {
-          name: "PRIMARY",
-          unique: true,
-          using: "BTREE",
-          fields: [{ name: "id" }],
-        },
-      ],
+    name: {
+      type: DataTypes.STRING(20),
+      allowNull: false,
+      unique: "name_UNIQUE"
     }
-  );
+  }, {
+    sequelize,
+    tableName: 'category_main',
+    timestamps: false,
+    indexes: [
+      {
+        name: "PRIMARY",
+        unique: true,
+        using: "BTREE",
+        fields: [
+          { name: "id" },
+        ]
+      },
+      {
+        name: "id_UNIQUE",
+        unique: true,
+        using: "BTREE",
+        fields: [
+          { name: "id" },
+        ]
+      },
+      {
+        name: "name_UNIQUE",
+        unique: true,
+        using: "BTREE",
+        fields: [
+          { name: "name" },
+        ]
+      },
+    ]
+  });
 };
