--- server/models2/tables/categoryMain.js
@@ -1,44 +1,0 @@
-module.exports = function (sequelize, DataTypes) {
-  return sequelize.define('categoryMain', {
-    id: {
-      type: DataTypes.INTEGER,
-      allowNull: false,
-      primaryKey: true
-    },
-    name: {
-      type: DataTypes.STRING(20),
-      allowNull: false,
-      unique: "name_UNIQUE"
-    }
-  }, {
-    sequelize,
-    tableName: 'category_main',
-    timestamps: false,
-    indexes: [
-      {
-        name: "PRIMARY",
-        unique: true,
-        using: "BTREE",
-        fields: [
-          { name: "id" },
-        ]
-      },
-      {
-        name: "id_UNIQUE",
-        unique: true,
-        using: "BTREE",
-        fields: [
-          { name: "id" },
-        ]
-      },
-      {
-        name: "name_UNIQUE",
-        unique: true,
-        using: "BTREE",
-        fields: [
-          { name: "name" },
-        ]
-      },
-    ]
-  });
-};
